requirements.txt
flask==2.3.3
requests==2.31.0
pyyaml==6.0.1
prometheus-client==0.17.1
redis==4.6.0
docker==6.1.3
psutil==5.9.5
schedule==1.2.0


